Sets up presentation layer structure
Organizes the project by introducing a presentation layer. This
includes adding `CardView` and `IconLabel` components to the design
system, refactoring the `CountriesListView` to use the new `CardView`
component, and creating a new `ContentView` to act as the main entry
point for the app's UI, which now uses a `NavigationStack`. The old
`ContentView` is replaced with a new one that displays a list of
items with enhanced UI elements.